Feat/new page validation (#3)
* feat: new page structure. Implementations and test adapted to

* feat: Notification struct changed. HTTPClient interface created to testing. Dependency Injection in validator service. Not tested yet

* feat: new mocks generated

* feat: dependency injection implemented for validator service. Notification struct moved to models package

* feat: validator test created

* fix: inline tag removed

* feat: new implementation consumed.

* feat: add some useful scripts

* feat: direct validatiion updated to distinguish a change in page

* feat: new interval to validate pages offline

+++ static/js/vue_impl.js
@@ -5,7 +5,6 @@ const app = new Vue({
         registerPageForm: {
             name: '',
             url: '',
-            status: true
         },
         registerPhraseForm: {
             phrase: ''
@@ -17,6 +16,7 @@ const app = new Vue({
         notifications_accepted: false,
         appClosed: false,
         evtSource: null,
+        validationInterval: null,
     },
     methods: {
         async requestNotificationsPermissions() {
@@ -39,8 +39,8 @@ const app = new Vue({
         async getPages() {
             const data = await fetch('/pages/all')
             this.pages = await data.json()
-            this.unactivedPages = this.pages.filter(i => i.status == false)
-            this.activedPages = this.pages.filter(i => i.status == true)
+            this.unactivedPages = this.pages.filter(i => i.is_working == false)
+            this.activedPages = this.pages.filter(i => i.is_working == true)
         },
         async getPhrases() {
             const data = await fetch('/phrases/all')
@@ -89,8 +89,8 @@ const app = new Vue({
                     return
                 }
                 
-                const statusPage = await res.json()
-                if(statusPage.error) {
+                const {page} = await res.json()
+                if(!page.is_working) {
                     alert("¡Listo!\nPágina registrada, pero no respondió")
                 }else {
                     alert("¡Listo!\nPágina registrada.")
@@ -194,6 +194,9 @@ const app = new Vue({
         async deletePage(data) {
             const onError = () => alert("ERROR AL BORRAR PAGINA.\nIntenta nuevamente.")
             try {
+                if(!confirm("¡Estas seguro en eliminar la página?")){
+                    return
+                }
                 const res = await fetch("/pages/delete", {
                     method: "DELETE",
                     body: JSON.stringify(data),
@@ -237,24 +240,25 @@ const app = new Vue({
         async sseHandler(e) {
             console.log("Creating notification")
             let data = JSON.parse(e.data)
-            let page = this.getPageByID(data.pageID)
-
-            if(data.recovered) {
+            if(!data) { // avoid process when there is no data in streaming
+                console.log("No data in strem")
+                return
+            }
+            let {page} = data
+            if(data.page.recovered) {
                 new Notification("¡Recuperada!", {
                     body: `La pagina '${page.name}' ya respondio :D`
                 })
-                await this.getPages()
-                return
+            } else if (!data.page.is_working) {
+                const phraseIndex = this.getRandomArbitrary()
+                new Notification(this.phrases[phraseIndex].phrase, {
+                    body: `La pagina '${page.name}' no responde`
+                })
+            } else if(data.page_id) {
+                new Notification(`Validación fallida`, {
+                    body: `Un error al validar la pagina '${page.name}'`
+                })
             }
-
-            if(!data.pageID) {
-                return
-            }
-            
-            const phraseIndex = this.getRandomArbitrary()
-            new Notification(this.phrases[phraseIndex].phrase, {
-                body: `La pagina '${page.name}' no responde`
-            })
             await this.getPages()
 
         },
@@ -275,21 +279,23 @@ const app = new Vue({
                     "Content-Type": "application/json"
                 }
             })
-            const statusPage = await res.json()
-            if(statusPage.recovered) {
-                alert(`La página ${page.name} fue recuperada :D`)
-                this.getPageByID(page.id).status = true
+            const {page:PageRes} = await res.json()
+            console.log(PageRes)
+            if(PageRes.recovered) {
+                alert(`La página '${PageRes.name}' fue recuperada :D`)
+                await this.getPages()
+                return
+            } else if (!PageRes.is_working){
+                alert(`La página '${PageRes.name}' aún no responde D:`)
+            }
+        },
+        setIntervalHandler(){
+            if(!this.unactivedPages.length > 0) {
                 return
             }
-            alert(`La página ${page.name} un no responde D:`)
-        }
-    },
-    computed: {
-        getUnactivedPages() {
-            return this.pages.filter(i => i.status == false)
-        },
-        getActivedPages() {
-            return this.pages.filter(i => i.status == true)
+            new Notification(`Recuerda...`, {
+                body: `Aún hay ${this.unactivedPages.length} página(s) sin responder :/`
+            })
         }
     },
     async created() {
@@ -301,5 +307,6 @@ const app = new Vue({
         this.evtSource.onmessage = this.sseHandler
         // this.evtSource.onopen = (e) => console.log("Connected to SSE")
         // this.evtSource.onerror = (e) => console.error(e)
+        this.validationInterval = setInterval(this.setIntervalHandler, 120000)
     }
 })
